Test additional variant payload types and split up to 1 file per method
